backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
claude-code-sdk>=0.0.25
python-dotenv>=1.0.0
pydantic>=2.0.0
